[Assessor] added ability to download original application PDF
[Trello Story](https://trello.com/c/TS7BZGtA/243-display-a-button-for-admins-to-download-the-original-pdf-of-the-application-at-the-deadline)

```
Assessors should also be able to download the orginal pdf
of the application at the deadline
```
