+++ app/controllers/admin/form_answers_controller.rb
@@ -16,18 +16,6 @@ class Admin::FormAnswersController < Admin::BaseController
     })
   end
 
-  expose(:original_form_answer) do
-    @form_answer.original_form_answer
-  end
-
-  expose(:pdf_data) do
-    original_form_answer.decorate.pdf_generator
-  end
-
-  expose(:pdf_filename) do
-    "original_pdf_before_deadline_#{@form_answer.decorate.pdf_filename}"
-  end
-
   def index
     params[:search] ||= FormAnswerSearch::DEFAULT_SEARCH
     authorize :form_answer, :index?
@@ -37,18 +25,6 @@ class Admin::FormAnswersController < Admin::BaseController
     @form_answers = @search.results.group("form_answers.id")
                                    .page(params[:page])
                                    .includes(:comments)
-  end
-
-  def original_pdf_before_deadline
-    authorize @form_answer, :can_download_original_pdf_of_application_before_deadline?
-
-    respond_to do |format|
-      format.pdf do
-        send_data pdf_data.render,
-                  filename: pdf_filename,
-                  type: "application/pdf"
-      end
-    end
   end
 
   private
